Fix local.properties value type.

+++ app/src/main/java/com/anhquan/pinotify/view/status/PlayIntegrityStatusViewModel.kt
@@ -62,7 +62,7 @@ class PlayIntegrityStatusViewModel : ViewModel() {
                 IntegrityManagerFactory.createStandard(context).prepareIntegrityToken(
                     PrepareIntegrityTokenRequest
                         .builder()
-                        .setCloudProjectNumber(BuildConfig.CLOUD_PROJECT_NUMBER)
+                        .setCloudProjectNumber(BuildConfig.CLOUD_PROJECT_NUMBER.toLong())
                         .build()
                 ).await()
         }
